% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-to-rprofile.R
\name{add_competition_to_rprofile}
\alias{add_competition_to_rprofile}
\title{Add a global variable to the .Rprofile file}
\usage{
add_competition_to_rprofile(competition)
}
\arguments{
\item{competition}{Name of the competition to be stored in the global variable.}
}
\description{
After calling this function, you will get a global variable named
\code{KAGGLE_COMPETITION} everytime you launch your  current R project.
}
