% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-to-rprofile.R
\name{add_pkg_to_rprofile}
\alias{add_pkg_to_rprofile}
\title{Add a package to the .Rprofile file}
\usage{
add_pkg_to_rprofile(pkg, global = FALSE)
}
\arguments{
\item{pkg}{Name of the package to be added to .Rprofile file.}

\item{global}{Whether to add it to the global .Rprofile (\code{TRUE}) or just
to the .Rprofile file of your current project (\code{FALSE}, the default).}
}
\description{
After calling this function, the package will be launched everytime you
launch your  current R project
(or any project if you use option \code{global = TRUE}).
}
\examples{
kaggler::add_pkg_to_rprofile("kaggler")
}
