% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle-api.R
\name{kaggle_submit}
\alias{kaggle_submit}
\title{Submit}
\usage{
kaggle_submit(file, message)
}
\arguments{
\item{file}{Submission file to upload.}

\item{message}{Message describing this submission.}
}
\value{
Input \code{file}, invisibly.
}
\description{
Submit a file to a Kaggle competition.
}
