% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggler-package.R
\docType{package}
\name{kaggler-package}
\alias{kaggler}
\alias{kaggler-package}
\title{kaggler: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
