% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{system_glue}
\alias{system_glue}
\title{Title}
\usage{
system_glue(...)
}
\arguments{
\item{...}{Arguments passed on to \code{glue::glue}
\describe{
  \item{.sep}{[\code{character(1)}: \sQuote{""}]\cr Separator used to separate elements.}
  \item{.envir}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored.}
  \item{.open}{[\code{character(1)}: \sQuote{{}]\cr The opening delimiter. Doubling the
full delimiter escapes it.}
  \item{.close}{[\code{character(1)}: \sQuote{}}]\cr The closing delimiter. Doubling the
full delimiter escapes it.}
}}
}
\description{
Title
}
