% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_csv}
\alias{to_csv}
\title{Write CSV}
\usage{
to_csv(x, file)
}
\arguments{
\item{x}{A named data frame.}

\item{file}{CSV file to write.}
}
\value{
Return input \code{file}, invisibly.
}
\description{
Write CSV file for submission.
}
