% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_files}
\alias{unzip_files}
\title{Unzip files}
\usage{
unzip_files(dir = "data", overwrite = FALSE)
}
\arguments{
\item{dir}{Directory where to get the files to unzip.}

\item{overwrite}{If \code{TRUE}, overwrite existing files, otherwise
    ignore such files.  The equivalent of \command{unzip -o}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Unzip all files in a directory.
}
