% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle-api.R
\name{kaggle_data}
\alias{kaggle_data}
\title{Download data}
\usage{
kaggle_data(dir = "data", cmd_add = "")
}
\arguments{
\item{dir}{Directory where to put these data (default is \code{"data"}).}

\item{cmd_add}{If you want to add something to the command.}
}
\description{
Download data of a Kaggle competition.
}
